[ExecuTorch] Add //examples/portable/executor_runner:executor_runner_opt
Attempt to add a Buck target that's analogous to the CMake build's executor_runner -- has all CPU ops that you need etc. The base executor_runner target is commented as intentionally having minimal deps, hence the separate target.

This is sort of a companion to https://github.com/pytorch/executorch/pull/9248, except that that PR is for CMake only and this PR is for Buck only.

Differential Revision: [D71220489](https://our.internmc.facebook.com/intern/diff/D71220489/)

ghstack-source-id: 271885480
Pull Request resolved: https://github.com/pytorch/executorch/pull/9291